Add links to MS translation engine
It seems there's no "right align" or anything similar in markdown.  This doesn't look very pretty but it works. 

Languages are sorted in country's English name.

Images are sourced from a repo located on github.com.

Feel free to reuse in another PR (if you want to put it in footer?) or edit (add more languages?).